Update to send shooter motor speeds to Dashboard
Removing string buffer and simply sending speeds to dashboard.  SmartDashboard will also update ShuffleBoard interface but you need to 
add widget to layout on driver station.  If you have a specific layout in mind, you can use Shuffleboard interface to put element in specific location on a specific tab on Shuffleboard but will not be available if you want to use SmartDashboard.